[Modify] route 수정 - 채팅방 페이지 메세지 보내기 컴포넌트 추가 - 마이프로필의 자식 페이지 추가 - 고정된 아래에 nav 띄우도록 수정(fixed) - route에서 navigation 배열 데이터 분리 - 프로필 route Outlet을 위한 레이아웃 파일 생성 - 프로필 경로 수정
