+++ src/layouts/ProfileRootLayout.jsx
@@ -1,0 +1,9 @@
+import { Outlet } from 'react-router-dom';
+
+export default function ProfileRootLayout() {
+  return (
+    <>
+      <Outlet />
+    </>
+  );
+}
