+++ src/pages/Profile/MyPage.jsx
@@ -3,8 +3,9 @@ import InfoNav from '@/components/MyPage/InfoNav';
 import MyMenu from '@/components/MyPage/MyMenu';
 import MyProfile from '@/components/MyPage/MyProfile';
 import Temperature from '@/components/MyPage/Temperature';
+import ProfileRootLayout from '@/layouts/ProfileRootLayout';
 
-export function Component() {
+export default function MyPage() {
   return (
     <div>
       <MyProfile />
@@ -12,6 +13,7 @@ export function Component() {
       <Temperature temp={41.2} />
       <ContentNav />
       <InfoNav />
+      <ProfileRootLayout />
     </div>
   );
 }
