+++ src/pages/index.js
@@ -16,3 +16,4 @@ export { default as ChatList } from './Chat/ChatList';
 export { default as UserProfile } from './Profile/UserProfile';
 export { default as EditProfile } from './Profile/EditProfile';
 export { default as ProfileDetail } from './Profile/ProfileDetail';
+export { default as MyPage } from './Profile/MyPage';
